Don't set PKGDATADIR to an absolute path
For we are going to check in a prebuilt bison to the source tree,
it makes no sense to embed a full absolute path in the bison binary.
That means you can run the built bison only from the top of the source
tree.
To run it from somewhere else, you need to set up environmental variable
BISON_PKGDATADIR.

Bug: 8517572
Change-Id: Iaa81b30781a6a9162c775cdadaa759dad79a179d